requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Monitoring
prometheus-fastapi-instrumentator>=6.1.0

# OpenAI
openai>=1.10.0

# Redis for caching
redis>=5.0.0

# Celery for background tasks
celery>=5.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0

# Telegram Bot
aiogram>=3.0.0

# Async file I/O
aiofiles>=23.2.0

# Security & sanitization
bleach>=6.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing extras
pytest-mock>=3.12.0
httpx>=0.27.0
